feat: added template
