+++ gb/public/scripts/index.mjs
@@ -1,0 +1,10 @@
+const data = await (await fetch("../data/index.json")).json();
+const areas = await (await fetch("../data/primary-area.json")).json();
+
+import { setSearchEventListeners, setSortingEventListeners, setUpAllPrimaryAreas } from "./table.mjs";
+
+setUpAllPrimaryAreas(data);
+
+setSortingEventListeners();
+
+setSearchEventListeners(areas, data);
